+++ R/gips_perm_class.R
@@ -28,10 +28,8 @@
 #' gperm <- gips_perm("(1,2)(5,4)", 7)
 #' gperm
 #'
-#' \donttest{
 #' try(gperm <- gips_perm(permutations::as.cycle("(5,4)"), 3))
 #' # Error, `size` equals 3 while the maximum element is 5.
-#' }
 #'
 #' @export
 gips_perm <- function(x, size) {
@@ -232,7 +230,7 @@ validate_gips_perm <- function(g) {
 #'
 #' @examples
 #' g_perm <- gips_perm(permutations::as.cycle("(5,4)"), 5)
-#' \donttest{print(g_perm)}
+#' print(g_perm)
 print.gips_perm <- function(x, ...) {
   validate_gips_perm(x)
   x <- permutations::as.cycle(x)